Missed vector type changes

+++ src/RegisterGroup.cpp
@@ -197,7 +197,7 @@ RegisterMapping::RegisterType RegisterGroup::getRegisterType() const
     return m_registerType;
 }
 
-uint16_t RegisterGroup::getStartingAddress() const
+int32_t RegisterGroup::getStartingAddress() const
 {
     return GroupUtility::getAddressFromString(m_mappings.begin()->first);
 }
@@ -270,7 +270,7 @@ void RegisterGroup::setDevice(const std::shared_ptr<ModbusDevice>& device)
     m_device = device;
 }
 
-uint16_t GroupUtility::getAddressFromString(const std::string& string)
+int32_t GroupUtility::getAddressFromString(const std::string& string)
 {
     auto firstAddressString = std::string(string);
     auto dotIndex = firstAddressString.find(SEPARATOR);
@@ -278,7 +278,7 @@ uint16_t GroupUtility::getAddressFromString(const std::string& string)
     {
         firstAddressString = firstAddressString.substr(0, dotIndex);
     }
-    return static_cast<uint16_t>(std::stoul(firstAddressString));
+    return static_cast<int32_t>(std::stoul(firstAddressString));
 }
 
 int16_t GroupUtility::getBitFromString(const std::string& string)
