+++ src/mappings/FloatMapping.h
@@ -38,7 +38,7 @@ public:
      * @param readRestricted Is the Mapping write only?
      * @param slaveAddress Slave address of device, leave to be assigned by device, default is -1.
      */
-    FloatMapping(const std::string& reference, RegisterType registerType, const std::vector<int16_t>& addresses,
+    FloatMapping(const std::string& reference, RegisterType registerType, const std::vector<int32_t>& addresses,
                  bool readRestricted = false, int16_t slaveAddress = -1);
 
     /**
